+++ app/controllers/submit.js
@@ -86,8 +86,8 @@ const MODAL_TITLE = {
 const SUBMIT_MESSAGES = {
     default: 'submit.body.submit.information.line1.default',
     moderation: 'submit.body.submit.information.line1.moderation',
-    [PRE_MODERATION]: 'submit.body.submit.information.line3.pre',
-    [POST_MODERATION]: 'submit.body.submit.information.line3.post',
+    [PRE_MODERATION]: 'submit.body.submit.information.line1.pre',
+    [POST_MODERATION]: 'submit.body.submit.information.line1.post',
 };
 
 const PERMISSION_MESSAGES = {
@@ -155,7 +155,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
 
     _State: State,
     // Project states - new project or existing project
-    applyLicense: false,
     newNode: false,
 
     // Information about the thing to be turned into a preprint
@@ -443,7 +442,7 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             ACTION.create.heading;
     }),
     buttonLabel: computed('moderationType', function() {
-        return this.get('moderationType') === PRE_MODERATION ?
+        return this.get('moderationType') ?
             ACTION.submit.button :
             ACTION.create.button;
     }),
@@ -453,7 +452,7 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             SUBMIT_MESSAGES.default;
     }),
     permissionInformation: computed('moderationType', function() {
-        return this.get('moderationType') === PRE_MODERATION ?
+        return this.get('moderationType') ?
             PERMISSION_MESSAGES.submit :
             PERMISSION_MESSAGES.create;
     }),
@@ -468,10 +467,17 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
         return !(state === ACCEPTED || (modType === POST_MODERATION && state === PENDING));
     }),
     editInformation1: computed('moderationType', function() {
-        return EDIT_MESSAGES.line1[this.get('moderationType')];
+        const moderationType = this.get('moderationType');
+        if (moderationType) {
+            return EDIT_MESSAGES.line1[moderationType];
+        }
     }),
     editInformation2: computed('moderationType', 'model.reviewsState', function() {
-        return EDIT_MESSAGES.line2[this.get('model.reviewsState')][this.get('moderationType')];
+        const reviewsState = this.get('model.reviewsState');
+        const moderationType = this.get('moderationType');
+        if (reviewsState && moderationType) {
+            return EDIT_MESSAGES.line2[reviewsState][moderationType];
+        }
     }),
     canResubmit: computed('moderationType', 'model.reviewsState', function() {
         const state = this.get('model.reviewsState');
@@ -668,7 +674,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             // when pressing 'Save and continue'.  Creates a preprint with
             // primaryFile, node, and provider fields populated.
             const model = this.get('model');
-            this.get('node.license').then(this._setDefaultPreprintLicense.bind(this));
 
             model.set('primaryFile', this.get('selectedFile'));
             model.set('node', this.get('node'));
@@ -776,7 +781,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
                 return;
             }
 
-            const node = this.get('node');
             const model = this.get('model');
 
             const copyrightHolders = this.get('basicsLicense.copyrightHolders')
@@ -786,19 +790,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             if (this.get('abstractChanged')) { model.set('description', this.get('basicsAbstract')); }
 
             if (this.get('tagsChanged')) { model.set('tags', this.get('basicsTags')); }
-
-            if (this.get('applyLicense')) {
-                if (node.get('nodeLicense.year') !== this.get('basicsLicense.year') || (node.get('nodeLicense.copyrightHolders') || []).join() !== copyrightHolders.join()) {
-                    node.set('nodeLicense', {
-                        year: this.get('basicsLicense.year'),
-                        copyright_holders: copyrightHolders,
-                    });
-                }
-
-                if (node.get('license.name') !== this.get('basicsLicense.licenseType.name')) {
-                    node.set('license', this.get('basicsLicense.licenseType'));
-                }
-            }
 
             if (this.get('doiChanged')) {
                 model.set('doi', this.get('basicsDOI') || null);
@@ -825,10 +816,9 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             }
 
             this.set('model', model);
-            this.set('node', node);
 
-            node.save()
-                .then(this._saveBasicsInfo.bind(this))
+            model.save()
+                .then(this._moveFromBasics.bind(this))
                 .catch(this._failMoveFromBasics.bind(this));
         },
 
@@ -1206,7 +1196,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
     _copyFile(copiedFile) {
         this.set('selectedFile', copiedFile);
         this.send('startPreprint', this.get('parentNode'));
-        this.set('applyLicense', true);
         this.set('newNode', true);
     },
 
@@ -1233,15 +1222,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
 
     _sendStartPreprint() {
         this.send('startPreprint');
-    },
-
-    _setDefaultPreprintLicense(license) {
-        // This is used to set the default applyLicense once a node is loaded,
-        // as if the node's license is not set or is of type No license,
-        // we want to set the default to make its license the same as the preprint license.
-        if (license === null || (license && license.get('name').includes('No license'))) {
-            this.set('applyLicense', true);
-        }
     },
 
     _finishUpload() {
@@ -1284,14 +1264,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             year: this.get('model.licenseRecord') ? this.get('model.licenseRecord').year : date.getUTCFullYear().toString(),
             copyrightHolders: this.get('model.licenseRecord') ? this.get('model.licenseRecord').copyright_holders.join(', ') : '',
         });
-    },
-
-    _saveBasicsInfo() {
-        const model = this.get('model');
-
-        model.save()
-            .then(this._moveFromBasics.bind(this))
-            .catch(this._failMoveFromBasics.bind(this));
     },
 
     _moveFromBasics() {
@@ -1429,7 +1401,6 @@ export default Controller.extend(Analytics, BasicsValidations, NodeActionsMixin,
             basicsLicense: null,
             subjectsList: A(),
             availableLicenses: A(),
-            applyLicense: false,
             newNode: false,
             attemptedSubmit: false,
         }));